+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/IssuerAgent.kt
@@ -17,6 +17,8 @@ import at.asitplus.wallet.lib.cbor.CoseHeaderCertificate
 import at.asitplus.wallet.lib.cbor.CoseHeaderNone
 import at.asitplus.wallet.lib.cbor.SignCose
 import at.asitplus.wallet.lib.cbor.SignCoseFun
+import at.asitplus.wallet.lib.data.CredentialToJsonConverter.toJsonElement
+import at.asitplus.wallet.lib.data.SdJwtConstants.SD_ALG
 import at.asitplus.wallet.lib.data.Status
 import at.asitplus.wallet.lib.data.VerifiableCredential
 import at.asitplus.wallet.lib.data.VerifiableCredentialJws
@@ -38,6 +40,8 @@ import kotlinx.serialization.json.buildJsonObject
 import kotlinx.serialization.json.encodeToJsonElement
 import kotlinx.serialization.json.jsonObject
 import kotlin.time.Clock
+import kotlin.time.Duration
+import kotlin.time.Duration.Companion.minutes
 import kotlin.time.Instant
 
 /**
@@ -48,6 +52,8 @@ class IssuerAgent(
     private val issuerCredentialStore: IssuerCredentialStore = InMemoryIssuerCredentialStore(),
     private val statusListBaseUrl: String = "https://wallet.a-sit.at/backend/credentials/status",
     private val clock: Clock = Clock.System,
+    /** Time to adjust the [Clock.now] for issuance date of credentials. */
+    private val issuanceOffset: Duration = (-3).minutes,
     override val cryptoAlgorithms: Set<SignatureAlgorithm> = setOf(keyMaterial.signatureAlgorithm),
     private val timePeriodProvider: TimePeriodProvider = FixedTimePeriodProvider,
     /** The identifier used in `issuer` properties of credentials (JWT VC and SD JWT). */
@@ -67,10 +73,11 @@ class IssuerAgent(
     override suspend fun issueCredential(
         credential: CredentialToBeIssued,
     ): KmmResult<Issuer.IssuedCredential> = catching {
+        val issuanceDate = clock.now().minus(issuanceOffset.absoluteValue)
         when (credential) {
-            is CredentialToBeIssued.Iso -> issueMdoc(credential, clock.now())
-            is CredentialToBeIssued.VcJwt -> issueVc(credential, clock.now())
-            is CredentialToBeIssued.VcSd -> issueVcSd(credential, clock.now())
+            is CredentialToBeIssued.Iso -> issueMdoc(credential, issuanceDate)
+            is CredentialToBeIssued.VcJwt -> issueVc(credential, issuanceDate)
+            is CredentialToBeIssued.VcSd -> issueVcSd(credential, issuanceDate)
         }
     }
 
@@ -204,6 +211,8 @@ class IssuerAgent(
             sdJwt.forEach {
                 put(it.key, it.value)
             }
+            // TODO make sure this only get's inserted at the top-level
+            put(SD_ALG, credential.sdAlgorithm.toIanaName().toJsonElement())
             vcSdJwtObject.forEach {
                 put(it.key, it.value)
             }